Navigation bug fixed and directions added

+++ SUPRating/Models/Business.swift
@@ -14,7 +14,7 @@ struct Business: Identifiable {
     let id = UUID()
     var title: String
     var subtitle: String
-    var rating: Double
+    var rating: Int
     var latitude: Double
     var longitude: Double
 }
